Refactor scalarized vector processing
This change replaces sequences of Extract/Insert for performing
operations on vectors one element at a time, with a helper function
which invokes a lambda expression or rr::Call() to a C function.

It makes use of a new element_count() function added to each Reactor
vector type. It uses snake_case instead of PascalCase to discern it
from intrinsics and other constructs as it is not meant to be an
integral part of the API surface.

Bug: b/214583550
Change-Id: I174c148714eff4925dd413077845205bf96e6eea