Construct SIMD pointers from an std::vector instead of array
Pointer4 is slated to become SIMD::Pointer and thus should not be
constructed from 4 scalar pointers or an std::array<> of size 4. This
change makes it take an std::vector<> instead so it can have arbitrary
SIMD width in a subsequent change.

The castFrom() methods were using 4 nullptr arguments to initialize a
Pointer4 variable as being of 'array' type. Instead of replacing that
with an std::vector<> of nullptr elements, these methods were replaced
with constructors which initialize the class to the right type.

Bug: b/214583550
Change-Id: I0eff17a9ff5ef4e0822e0108a6aad64727c81249