Omit dynamic checks for equal or sequential pointer offsets
Previously we would dynamically check whether the offsets of a SIMD
pointer are equal or sequential, in which case gather and scatter
operations can be performed with a single scalar or packed load or
store. However, this check itself is expensive and it also requires
a dynamic branch operation with bad predictability. It also does not
scale well to wider SIMD.

Instead we should take advantage of dynamic uniformity when indicated
by the SPIR-V attributes. Also, it is expected that modern CPUs handle
gather and scatter instructions with greater efficiency if the offsets
are equal or sequential.

Bug: b/236280746
Change-Id: I852a45daf8d7fe2e97931d68f6ae37004ece7edc