Temporarily support xyzw component swizzling for SIMD types
Several places in the code currently make use of component swizzling,
even for code which should be SIMD width agnostic. To facilitate
transition to rr::SIMD types, this change temporarily adds support for
these operators.

Note that Swizzle(SIMD::*) has asserts for SIMD::Width == 4 so we get
reminded to remove this later.

Bug: b/214583550
Change-Id: I6ee5ea3a668eb63cab8ed0763c148ce4bbd90110