Disable AddressSanitizer JIT instrumentation support on Windows
When compiling with Clang on Windows, __has_feature(address_sanitizer)
is true for asan enabled builds. This caused us to try to include
<dlfcn.h> and use dlsym(), which is UNIX specific, for looking up
symbols in the asan runtime for use by the LLVM JIT instrumentation.

While it should be feasible to do the equivalent on Windows by using
GetProcAddress(), for now just disable asan instrumentation of JIT-
compiled code on Windows to fix the asan Clang build.

Instrumentation of static code (C++) is unaffected.

Bug: b/240465596
Change-Id: Ibeee6dcdc233c9d43dd0dfd30ac6c8c12ec3711d