Compute fragment coordinates for wider SIMD groups
This change removes precomputing a 'quad' of the primitive's
origin coordinates, instead just storing the coordinates of the primary
vertex. For point primitives this corresponds to the PointCoord builtin.

Fragments are laid out within a SIMD group as a Z-order curve, also
known as Morton order. The new compactEvenBits() function assists with
computing the relative x and y coordinates of each fragment and is
based on code by Fabian Giesen: https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/

Note this change removes work from the setup routine, but adds the minor
cost of two SIMD additions per subgroup in the pixel routine. If the JIT
compiler doesn't already optimize this through loop-invariant code
motion, we could precompute it ourselves once per rasterized primitive.

Bug: b/237494823
Change-Id: I9f0956e5937a4c8687ba7461291ab8a7a9d67307