Simplify clipping of complex polygons
Polygons constructed for point and line rendering previously had clip
flags computed for each new vertex. This costs about as much as actually
performing full polygon clipping, so we can just do that immediately.

Note that for triangles we still take advantage of the clip flags that
cheaply get computed in parallel in the vertex routine, to skip polygon
clipping when fully inside the frustum.

Bug: b/207658159
Change-Id: I17e6c95bb7e9ffe601f5e4403cb0c43c62f7f941