VK_KHR_global_priority implementation
VK_KHR_global_priority is the only required extension of
roadmap 2022. See:
https://registry.khronos.org/vulkan/specs/1.3-extensions/html/vkspec.html#_required_extensions

This CL implements the VK_KHR_global_priority extension,
and exposes the following features:
- globalPriorityQuery is set to true
- 2 possible queue priorities are exposed (low and medium)

This is the minimal subset of features I found would make
all global priority related tests pass simultaneously. I
opened an issue in dEQP-VK to verify if this is expected
or if this is a test bug:
https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/3898

Note that the current implementation is noop (low and medium
priorities are treated as the same priority).

Tests: dEQP-VK.api.*global_priority*
Bug: b/243049774
Change-Id: Ifa9d02b82d0568ed7cfcd5b2a3bc988debb3223f