Clarify negativeOneToOne depth clipping comments
While VK_EXT_depth_clip_control affects both the viewport transform and
primitive clipping, the latter is not concerned with how the z
coordinate gets mapped between the near and far frustum plane. Therefore
to avoid confusion the comment which mentioned such mapping has been
removed. For the clipping code we only need to know that the position of
the near clipping plane changes.

Bug: b/207658159
Change-Id: I40fd1c82c85a1841de91d22f137d8fa925ec0ada