Bogus VK_EXT_multisampled_render_to_single_sampled
DO NOT SUBMIT

This is a hack to test ANGLE's use of
VK_EXT_multisampled_render_to_single_sampled together with VVL.  Instead
of actually rendering multisampled, this implementation turns
multisampled pipelines into single sampled.

Fails tests in ANGLE that mix MSRTSS and multisampled attachments, and
another that specifically ensures multisampled rendering is done with
multiple samples.  However, the rest of the tests pass as they are not
sensitive to rendering being actually multisampled.

Discovered a spec bug:
https://gitlab.khronos.org/vulkan/vulkan/-/merge_requests/5392

Change-Id: If036a9a1350dbd6b6f8a4e975dc1d60e1e6e1a4f