shaderSampledImageArrayNonUniformIndexing support
This CL implements the shaderSampledImageArrayNonUniformIndexing
feature, which allows arrays of samplers or sampled images to be
indexed by non-uniform integer expressions in shader code.

This is a brute force implementation, which simply repeats the
sampling operation for every lane. This leaves room for future
optimization when some of them or all of them don't diverge.

Bug: b/236957718
Change-Id: Ic4d8b58b1bd8281a07c28a43d9a96f3ae2ea9bf4