Fix Frexp() for subnormal arguments under DAZ
Previously only an all-zero exponent and mantissa field was considered
equal to zero. When denormals-are-zero mode (DAZ) is enabled, currently
the default, we should treat subnormals as zero as well. This is
accomplished using a floating- point compare operation instead of
integer comparison.

Note that, just like the previous implementation, this still does not
handle subnormals correctly when DAZ mode is not enabled. To return a
significand in the range [0.5, 1.0) it will have to support returning
an exponent smaller than -126.

Use of the zeroSign variable was eliminated by observing that we already
obtained the sign bit but cleared it when isNotZero is false, which can
be avoided by instead inserting the exponent for 0.5 only when the
argument is not zero.

The Exponent() helper function did not return the actual exponent of an
IEEE 754 single-precision value. The exponent bias is 127, not 126, but
the latter is correct for implementing frexp() since the significant is
in the [0.5, 1.0) range while mantissas represent [1.0, 2.0). So the
Exponent() function was eliminated and its logic inlined in Frexp().

Bug: b/243791551
Bug: b/169904252
Change-Id: I67b7a5a8e971a0290aa5cedb55acae6e70f9485c