Fix Frexp() for subnormal arguments under DAZ
Previously only an all-zero exponent and mantissa field was considered
equal to zero. When denormals-are-zero mode (DAZ) is enabled, currently
the default, we should treat subnormals as zero as well. This is
accomplished using a floating- point compare operation instead of
integer comparison.

Note that, just like the previous implementation, this still does not
handle subnormals correctly when DAZ mode is not enabled. To return a
significand in the range [0.5, 1.0) it will have to support returning
an exponent smaller than -126.

Bug: b/243791551
Bug: b/169904252
Change-Id: I67b7a5a8e971a0290aa5cedb55acae6e70f9485c