Eliminate incorrect Exponent() function
The Exponent() helper function did not return the actual exponent of an
IEEE 754 single-precision value. The exponent bias is 127, not 126, but
the latter is correct for implementing frexp() since the significant is
in the [0.5, 1.0) range while mantissas represent [1.0, 2.0). So the
Exponent() function was eliminated and its logic inlined in Frexp().

Bug: b/243791551
Change-Id: I18eaa6930c4583468e6129126d64227f30fd9696