Add a reference implementation for frexp()
sw::Frexp() currently only works correctly when the CPU treats
subnormals as zero. This satisfies the corresponding SPIR-V shader
instruction requirements.

This change demonstrates a C++ reference implementation of frexp() which
also supports subnormal values, in case we have a future need for it.
It works by multiplying subnormal values by a power of 2 which makes
them normalized. This adds to the value's exponent. The exponent of the
normalizing factor is subtracted again when obtaining the exponent of
the normalized value.

The normalizing factor must larger than or equal to 2^23 for the
smallest subnormal value, but must be no larger than 1.0 for the largest
non-infinite value to avoid overflow. So it is derived from the lower
7 bits of the input value's exponent.

Infinity and NaN are also handled in accordance with the C++ <cmath>
specification for std::frexp().

Bug: b/243791551
Change-Id: I68bcfa06379fc858acf8ccf7634cb7812e1dd1cd