Initial support for VK_EXT_pipeline_robustness
In this initial implementation, enabling robustBufferAccess on any of
storage, uniform or vertex buffers makes it enabled for all of them.

This can be optimized in a future change, by appropriately tracking
separate robustBufferAccess flags for storage, uniform and vertex
buffers and using the right flag for the right buffer.

Bug: b/185122256
Tests: dEQP-VK.robustness.pipeline_robustness*
Change-Id: Idb9f841e2a850b866cdfbc1d558efb2d33203c4d