OpSampledImage propagation fix
This CL fixes issues when the result of an OpSampledImage instruction
is copied. We can no longer rely on the following instructions being
able to query if the previous instruction was OpSampledImage, since
it's also acceptable to use copy operations on the result of
OpSampledImage operations.

If a sampler is associated with an image, either when performing an
OpImage, OpSampledImage or OpCopy* operation, the sampler is now
properly propagated and queried in the ImageInstruction constructor
in order to retrieve the proper sampler, even if it doesn't come
directly from an OpSampledImage instruction.

Bug: b/236957718
Change-Id: I7e6a7834f955fb0be5a50f4f5b1717a471a7aaf6