Skip Setup and Pixel JIT if rasterizer discard
Split from the VK_EXT_graphics_pipeline_library work.  This change makes
sure that if rasterizer discard is enabled, the Setup and Pixel steps
are not only not performed, but neither updated nor JITed at all.

The change helps fail eariler through a crash (as opposed to reading
unset state) if these steps are accidentally used with rasterizer
discard.  Additionally, it's also a small optimization for pipelines
that are never used without rasterizer discard.  Most importantly
however, it's a code structuring change to facilitate future changes
with partial pipeline JIT.

Additionally, to support this change, rasterizerDiscard is moved from
SetupProcessor::States to DrawData, so that the setupState wouldn't need
updating.

Bug: b/245568070
Tests: dEQP-VK.*
Change-Id: I8efadd07efa46a3e82d86c7d3592ef27282d5678