Skip Setup and Pixel JIT if rasterizer discard
Split from the VK_EXT_graphics_pipeline_library work.  This change makes
sure that if rasterizer discard is enabled, the Setup and Pixel steps
are not only not performed, but neither updated nor JITed at all.

The primary purpose of the change is to ensure these steps are not
accidentally used with rasterizer discard.  However, this change is also
a small optimization for pipelines that are never used without
rasterizer discard.

A few additional minor changes are made in this change:

- rasterizerDiscard is moved from SetupProcessor::States to DrawData, so
  that the setupState wouldn't need updating
- A hack to set sample count to 1 is moved from GraphicsState's
  constructor to Renderer::draw itself.  This is where batch size is
  derived from sample count.  A side effect of this change is that the
  batch size is now based on sample count == 1 on an otherwise
  multisampled pipeline that has rasterizer discard dynamically enabled.

Bug: b/245568070
Tests: dEQP-VK.*
Change-Id: I8efadd07efa46a3e82d86c7d3592ef27282d5678