Move sample count hack from GraphicsState to Renderer::draw
A hack to set sample count to 1 is moved from GraphicsState's
constructor to Renderer::draw itself.  This is where batch size is
derived from sample count.  A side effect of this change is that the
batch size is now based on sample count == 1 on an otherwise
multisampled pipeline that has rasterizer discard dynamically enabled.

Bug: b/245568070
Tests: dEQP-VK.*
Change-Id: I9f2ddfd9b0a7f94eca417d8a521ed6f7d793846b