Temporary patch for OOB read issue
There's an open VK-GL-CTS issue:
https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/3974
regarding tests which cause SwiftShader to perform an OOB read
operation. There's also a proposed fix:
https://gerrit.khronos.org/c/vk-gl-cts/+/10260
for this issue. Since this is causing an OOB read issue in
SwiftShader by performing a buffer to image copy from a buffer
which is too small, there's no possible workaround for it in
SwiftShader, so, for now, we temporarily apply this patch until
the issue is properly resolved in VK-GL-CTS.

Tests: dEQP-VK.spirv_assembly.instruction.graphics.image_sampler.imagesample_dref_*optypeimage_mismatch*
Bug: b/247637093
Change-Id: I66fc9cdf2b592a45e740787f192f6924b711e0a7