Partially revert "Pipeline robustness fix"
This CL partially reverts:
https://swiftshader-review.googlesource.com/c/SwiftShader/+/68331

Some of the changes were unnecessary and caused
VK_PIPELINE_ROBUSTNESS_BUFFER_BEHAVIOR_DEVICE_DEFAULT_EXT to no
longer be properly supported when specified. This was not uncovered
because of the lack of test coverage. This CL fixes the issue by
setting inheritRobustBufferAccess to deviceRobustBufferAccess when
overrideRobustness IS specified, instead of when it IS NOT specified,
which is what the original issue was.

Tests: dEQP-VK.robustness.pipeline_robustness_buffer_access.*
Bug: b/185122256
Change-Id: I9c09b9dc3a25dacd35e166189eb828f78dd9c422