Avoid inadvertently adding a wayland dependency
This BUILD.gn adds libWaylandClient.[cpp,h] to the sources on linux
when using the ozone wayland backend, but adds a dependency on
libwayland-client regardless of the ozone backend.

This patch avoids adding the dependency when not building the
wayland ozone backend.

This fixes #177 (https://bugs.chromium.org/p/swiftshader/issues/detail?id=177)

Change-Id: Ia1d58d200d1059e9737933431b3f9334ca545653