Remove incorrect assert
An assert in Framebuffer::setAttachment() was checking that the
attachment was null before it being assigned, but there is no
such requirement in the Vulkan spec. The assert was removed.

Also fixed an issue where the attachments were set at command
recording time, for some reason, instead of command execution time.

A corresponding issue was logged in VK-GL-CTS:
https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/3999

Tests: dEQP-VK.imageless_framebuffer.*
Bug: b/248611054
Change-Id: Ibd74dce59a315df0cfb25879d9e495f675f4b01c