Consistently format pointer-to-const
This change replaces occurrences of `T const *p` with `const T *p`.

The latter was already the predominant style. While consistency is a
strong motivation for this change in itself, this arguably also
improves readability.

The common argument for the former order of the const qualifier is that
compilers primarily perform type parsing in right-to-left order. While
there are cases where we need to pay special attention to that, types
like `unsigned int *` read more naturally as a pointer to an unsigned
integer than `int unsigned *`, i.e. we mentally group the modifier with
the basic type and read this part left-to-right.

This also avoids confusion with `int *const` which means something else
than `int const *`. Writing the latter as `const int *` makes them
easier to discern.

`constexpr int *p` actually means `p` is (compile-time) constant, not
its pointee. To make the pointee constant also one would need
`constexpr const int *p`. Since this has the appearance of redundancy
this change formats it as `const int constexpr *p` to bring the
`constexpr` closer to `p` and make it look more similar to
`const int *const p`.

Bug: code style formatting
Change-Id: If6ab49ecbdab8037d46df7581358129de5577902