Limit SPIR-V phi variable lifetimes
This change moves phi variables from SpirvRoutine to EmitState. The
former class represents the interface of a shader, and phis are not part
of that. EmitState only lives during emission of the actual SPIR-V code
into Reactor code, so it deletes these variables automatically and
prevents unnecessary materialization and reduces compilation overhead.

This also enabled reducing the public interface of EmitState to a single
static emit() function.

Bug: b/178662288
Bug: b/247020580
Change-Id: I88110c02e925699ad9706aaa3dc8779f9bbd3430