Unplug the shader debugging support
It is no longer maintained and has bitrotted to the point where
upcoming refactorings are hindered by it and we're likely to break it
even more in ways that will just add confusion if/when we revive it.

This change unplugs it entirely so we can re-architect active parts of
the code and bring shader debugging support back after the dust settles
and we have a need for it again. Note the bulk of the implementation is
in SpirvShaderDebugger.cpp and is kept so that we have a useful code
history, assuming we'll base the future version of it on the same code.

Bug: b/251802301
Change-Id: I0f30a6b0a61a0ff5a9c05a6be24f4f1935f84699