Unplug SPIR-V instruction profiling support
Similar to the shader debugging support this feature is not actively
used and there are no unit tests to prevent it from bitrotting. It gets
in the way of refactorings which will make the SpirvShader class solely
responsible for SPIR-V parsing and independent of Reactor code emit.

It should be straightforward to revive if we need it again and make an
effort to properly maintain it.

Bug: b/251802301
Change-Id: I5f54fa937bd5461057afa8f4d8d61679ac642f4d