Refactor GroupNonUniformArithmetic handling
The implementation of GroupNonUniformArithmetic instructions used to
rely on a pimpl struct containing a static template function. It's
not very elegant and causes code bloat. It's also too closely tied to
the SpirvShader class, which we want to refactor into a parsing-only
class.

This change reduces the template function to performing the operation
on scalarized components. It doesn't need access to SpirvShader and
takes one fewer arguments.

Bug: b/247020580
Change-Id: I24955b42d84a3a31f139bc0a3aacc09e82f58fee