Rename SpirvShader to Spirv
SpirvShader currently holds both the parsed SPIR-V code and pipeline
state which will affect code emission. This state is provided at
pipeline creation time but we perform code emission at draw time.

This change essentially renames SpirvShader to Spirv, in preparation of
making the latter a purely parsing-only class, and a new SpirvShader
class was added which derives from Spirv and is meant to store the
state variables in a subsequent refactoring.

Bug: b/247020580
Bug: b/253234336
Change-Id: I582353a587d5911a05f6cbb3129f18f27f23ca20