Eliminate EmitResult
The return value of code emission functions was used to determine
whether the terminator of a SPIR-V code block had been reached. This was
not intuitive and entangled a control flow implementation detail with
all unrelated instructions as well.

Arguably one advantage of the old code is that it was impossible to not
add an EmitResult to the implementation of new instructions, while the
new IsTerminator() function returns false in the default case. To
prevent this from causing bugs when left unupdated, the
EmitInstruction() method now handles terminator and non-terminator
instructions in separate switch statements.

Bug: b/247020580
Change-Id: I8ea1fa2250e22990f0f4864d1844c8e378793f1e