Fix zlib dependency for Regres
VK-GL-CTS has updated the zlib dependency, because the previous one
no longer exists. This causes Regres to fail to fetch the proper
version of zlib and exit early. This updates the hash to the current
ToT, where the only difference from the previous version we were
using is the zlib version fix, so there should be no test updates.

Bug: b/201453797
Change-Id: I50857e3ca70203e383d1c0d4ed5b98b2d0958e60