Use sw::bit_cast<> for type punning
Type punning through pointer or reference type cast breaks strict
aliasing rules which can lead to undefined behavior.

Note that sw::bit_cast<> makes use of a union to perform type punning,
which is also in violation of strict C++ rules, but is supported by
extension in at least the Visual Studio, Clang, and GCC compilers. It
will also be trivial to replace with std::bit_cast<> once we can upgrade
to C++20.

Bug: b/227975743
Change-Id: I4b94764aa6445d7c8d6463ba3a4b23a09ac975ef