Default to masking stores in helper invocations
The Vulkan spec states that "Stores and atomics performed by helper
invocations must not have any effect on memory except for the Function
and Private storage classes". Thus the switch statement which checks the
storage class of stores in helper invocations must default to not
causing any effect.

The Function and Private storage classes are the exception, and the
discussion at https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/4060
strongly supports that the Output storage class should get the same
treatment.

Note that this code previously also considered Output to have effect, so
this change is the closest thing to a refactoring and we can alter the
behaviour later if deemed incorrect after all.

The method was renamed and meaning inverted for additional clarity of
its semantics.

Bug: b/253701784
Change-Id: I0315d3e6ed4571307f1d6d2c5a33bff5d6f096d0