Move SIMD functions to LLVM and Subzero implementations
This CL splits the implementations of functions which require
looping over groups of 4 SIMD elements into Subzero and LLVM
implementations. The Subzero implementation only supports a SIMD
width of 4, so the loops are removed. The LLVM implementation is
unchanged from the SIMD implementation, but the loops will be
removed as proper implementations of these functions are done.

Bug: b/237494823
Change-Id: Ife15d1c6d8cba05b9bdbfa49678b62500832b0dd