Support larger SIMD widths with Rcp, RcpSqrt and Sqrt
This CL introduces a new function in LLVMReactor.cpp, called
GetIntrinsic(), which will be used to find the proper intrinsic,
based on the desired operation and the SIMD::Width.

As an example, this was used to change a few functions (Rcp,
RcpSqrt and Sqrt), so that they support larger SIMD widths by
changing which intrinsic is selected based on the SIMD::Width.

This should allow us to easily make other similar operations
support larger SIMD widths.

Still todo: Unit tests will need to support SIMD::Float better.

Bug: b/237494823
Change-Id: I596a11461ceeeef89bc6d10d171ad3c9e4d88e8b