Float only writeColor
This CL remove the version of writeColor() using short inputs,
as the version with floating point inputs now supports all
formats which were supported by this function.

The linearToSRGB16_12_16() function is also eliminated as a
side effect of this change.

Finally, this CL also removes SWIFTSHADER_LEGACY_PRECISION,
since both this change and SWIFTSHADER_LEGACY_PRECISION may
cause the need for rebaselines, so we combine these in a single
CL to minimize rebaselining.

Bug: b/204322086
Bug: chromium:1299047
Change-Id: I074e2f9dffbdc54f927ab73f563543ab6d83d29b