Fix timeline semaphore wait criteria
If VK_SEMAPHORE_WAIT_ANY_BIT is set, the current behavior is to avoid
waiting and the semaphore's current value is exactly the value waited
for.

If VK_SEMAPHORE_WAIT_ANY_BIT is not set but there is a non-infinte
timeout, the current behavior is to call wait_until with the criterion
that the counter reaches the exact value waited for.

In both cases, larger values should also be accepted.

Change-Id: I8e447501cc1d09d2071c95398663ec06b444106d
Bug: b/265933001