Fix vkWaitSemaphores when VK_SEMAPHORE_WAIT_ANY_BIT
The current implementation of TimelineSemaphore has a number of problems:
1. Shared::signal only works properly when the value signaled is
   exactly the same as the value waited for.
2. Dependents added to a semaphore are never cleaned up, so each
   wait-for-any operation on a semaphore adds one more dependent. This
   leads to long signaling times after a few thousand waits.
3. There's a race between checking for the current value of a timeline
   semaphore and adding the dependent to the 'any' semaphore. This
   causes waits to sometimes hang indefinitely.

Since the Shared part of TimelineSemaphore seems to exist solely to
support wait-for-any, this CL removes it in favor of a simpler
implementation without shared_ptrs. That behavior is instead
implemented by the WaitForAny object, which is intended to live on the
stack and which will clean up after itself during destruction.

Change-Id: Ie82c0a5520708c6dda90f594ffdd467a0d927a75
Bug: b/265933001
Tests: dEQP-VK.synchronization.basic.timeline_semaphore.*