Fix out-of-bounds operand access in image write.
SwiftShader incorrectly accessed nonexistant components
of the texel operand of an ImageWrite instruction. For example,
even in the case of a 1D image write (where the texel
has 1 component), SwiftShader would still emit code to access
4 components.

Despite the components not being used in the final JITted code,
accessing the data structure holding the operand's components
at compile time resulted in out-of-bounds accesses.

I have also added assertions that accesses into an operand's
components are always in-bounds (assertions already existed
when the operand was an intermediate but not when it was a
constant).

Bug: None

Change-Id: I85ba79b18ef82f1ef676aafe2639a289372e254b