[LLVM16] Fix crash when findLSB is called with 0.
With LLVM 16, dEQP-VK.graphicsfuzz.cov-inc-array-element-loop-lsb
started to crash.

It appears that Swiftshader was generating a Cttz call
(https://llvm.org/docs/LangRef.html#llvm-cttz-intrinsic)
with <is_zero_poison> set to true, even though the parameter
can be zero. This led to undefined behavior.

With this change, the test does not crash anymore.

b/272710814

Change-Id: I912a770ee808a54a59eca1da45b41636708e80bb