[LLVM16] Fix crash when findLSB is called with 0.
With LLVM 16, dEQP-VK.graphicsfuzz.cov-inc-array-element-loop-lsb
started to crash.

It appears that Swiftshader was generating a Cttz call
(https://llvm.org/docs/LangRef.html#llvm-cttz-intrinsic)
with <is_zero_poison> set to true, even though the parameter
can be zero. This led to undefined behavior.

With the change in SpirvShaderGLSLstd450.cpp, the test does not
crash anymore.

While searching for more instances of this bug, the call in
SpirvShaderGroup.cpp was found.
The change in SpirvShaderGroup.cpp is unrelated to the dEQP failure.

b/272710814

Change-Id: I912a770ee808a54a59eca1da45b41636708e80bb