[LLVM 16] bitfieldExtract handles negative offset.
This CL addresses the test failures that occurs in
dEQP-VK.graphicsfuzz#cov-bitfieldExtract-undefined
when Swiftshader uses LLVM 16.

This test verifies that you can pass a negative
offset to `bitfieldExtract` without causing UB.
This test fails when Swiftshader uses LLVM 16.

This is because Swiftshader implements
`bitfieldExtract` using LLVM's lshr instruction,
which causes a poison value to be generated if
the shift is larger than the number of bits.
See https://llvm.org/docs/LangRef.html#lshr-instruction

This CL takes advantage of LLVM's freeze IR instruction
to absorb the poison value.

Test: Ran the DEQP-VK tests.
Change-Id: I32e4afe2ba9b993942d1eb1cd9b9d339533c98a8