[LLVM 16] bitfieldExtract handles negative offset.
This CL addresses the test failures that occurs in
dEQP-VK.graphicsfuzz#cov-bitfieldExtract-undefined
when Swiftshader uses LLVM 16.

This test verifies that you can pass a negative
offset to `bitfieldExtract` without causing UB.
This test fails when Swiftshader uses LLVM 16.

This is because Swiftshader implements
`bitfieldExtract` using LLVM's lshr instruction,
which causes a poison value to be generated if
the shift is larger than the number of bits.
See https://llvm.org/docs/LangRef.html#lshr-instruction
For the record, compiling `a >> -1` on godbolt
with `-emit-llvm -O3` shows that this behavior
started happening with clang 12.

It is possible to persuade llvm into not generating a poison value
even if `rhs` is too large by creating the following IR:
```
    if rhs >= 32
       return undef
    else
       return lshr(lhs, rhs)
```

The optimizer will realize that instead of returning `undef`
(an undefined value), it might just as well always return `lshr`.
You can verify this by compiling with -O3 the following LLVM IR:
```
    define i32 @shiftbynonconst(i32 %a, i32 %shift) local_unnamed_addr #0 {
          %ok = icmp slt i32 %shift, 32
          %shifted = shl i32 %a, %shift
          %result = select i1 %ok, i32 %shifted, i32 undef
          ret i32 %result
    }
```

and noticing that the `select` (equivalent to the `if`) is optimized away:
```
    shiftbynonconst:                        # @shiftbynonconst
            mov     ecx, esi
            mov     eax, edi
            shl     eax, cl
            ret
```

The SIMD variant of the IR behave similarly.

Change-Id: I32e4afe2ba9b993942d1eb1cd9b9d339533c98a8