Support version 7 of the ICD interface.
Reland after making the version 5 requirement only apply if the Vulkan
loader loaded Swiftshader (and not when an application, like Chromium,
loads vkGetInstanceProcAddr directly).

The Vulkan loader requires this version for direct driver loading, which
Chromium / Dawn would like to use for loading Swiftshader with the
Vulkan loader (so that VVLs can be inserted).

 - Version 4 adds vk_icdGetPhysicalDeviceProcAddr.
 - Version 5 adds a check for the apiVersion depending on the negotiated
   ICD interface version.
 - Version 6 adds vk_icdEnumerateAdapterPhysicalDevices on Windows
   (which is a noop for Swiftshader)
 - Version 7 doesn't require any code change.

Bug: dawn:2145
Change-Id: Idb5b7bd473a13f5ab3c2319174f80a220599b237