Fix unaligned_ptr in swiftshader
For something used in three places, this type is probably overkill, but
ah well. Casting an unaligned pointer to uint32_t* is UB, even if you
don't dereference it. So this unaligned_ptr type isn't actually avoiding
UB. Keeping things as void* under the hood, to avoid this.

While I'm here, trim the API surface a bit, so this is a bit easier to
reason about.

Bug: chromium:1394755
Change-Id: I3a198c78051508c6b2ec806ce0ea6b4335cf4ebe