Restore Bug line in commit message requirement
While it's a minor convenience to not require a Bug line in the commit
message to pass Kokoro tests, it can lead to major inconveniences to try
to understand the exact motivation behind changes and have an overview
of all changes related to a common topic (which may land far apart).
Clarifying information cannot be added to commit messages after they've
landed, but bug trackers allow for discussions and documentation to be
recorded both before and after changes land. Having Kokoro produce an
error if a properly formatted Bug line is missing helps ensure these
productivity and quality enhancing benefits are retained.

This reverts commit 85772f1891dbbab714319980ba1f26c562f0282e.

It argued that most changes are now community contributions, but only
4 out of the last 100 changes were contributed by non-Googlers. Also,
only 10 changes did not have a Bug line, which illustrates that most
people are used to adding a bug reference and deem it valuable.

Lastly, the presubmit check does not strictly demand a reference to a
bug tracker. It also accepts any string not containing numbers, thus a
description like "Bug: comment spelling fix" or even "Bug: none"
technically suffices and avoids requiring opening a tracker bug for
arguably minor standalone changes.

Bug: b/141892461
Change-Id: I71f9e85ef9c81694d552628ada93afdcb3b639de