Fix some more UBSan errors in swiftshader
VkPipelineLayout.cpp had a memcpy call that could be passed NULL, 0.
Switch to std::copy which avoids this and also a manual sizeof
multiplication.

VkStructConversion.cpp is a bit more fun. It manually lays out a bunch
of arrays in a single structure. Almost all fields in the array provide
8-byte alignment, except for VkPipelineStageFlags. If there are an odd
number of VkPipelineStageFlags, subsequent fields become misaligned.
This is undefined not only by C/C++ (it is forbidden to cast unaligned
pointers), but also Vulkan itself if I'm understanding the spec right:
https://registry.khronos.org/vulkan/specs/1.3-extensions/html/vkspec.html#fundamentals-validusage-pointers

(Although Vulkan saying it is somewhat redundant because merely having
those pointer types in the type signature means you must keep them
aligned by C/C++'s rules.)

Fix this by rounding all the array sizes up to a multiple of 8. This is
a no-op for everything but the VkPipelineStageFlags fields, but I
applied it to all fields so it is more obviously correct. The compiler
is smart enough to optimize this out per [0].

[0] https://godbolt.org/z/nfM3Y6xfo

Bug: chromium:1394755
Change-Id: I950869e4fed969351ac1aefddd72a898b43cfdc1